refactor : review 도메인 분리
