+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/review/dto/ReviewRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.review.dto;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
