+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/review/exception/ReviewExceptionMessage.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.exception;
+package com.sparta.blackwhitedeliverydriver.review.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
