+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/review/service/ReviewService.java
@@ -1,21 +1,21 @@
-package com.sparta.blackwhitedeliverydriver.service;
+package com.sparta.blackwhitedeliverydriver.review.service;
 
-import com.sparta.blackwhitedeliverydriver.dto.ReviewIdResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.ReviewRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.ReviewResponseDto;
-import com.sparta.blackwhitedeliverydriver.entity.Order;
-import com.sparta.blackwhitedeliverydriver.entity.Review;
-import com.sparta.blackwhitedeliverydriver.entity.Store;
-import com.sparta.blackwhitedeliverydriver.entity.User;
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import com.sparta.blackwhitedeliverydriver.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.OrderExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.ReviewExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.StoreExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.repository.ReviewRepository;
-import com.sparta.blackwhitedeliverydriver.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.review.dto.ReviewIdResponseDto;
+import com.sparta.blackwhitedeliverydriver.review.dto.ReviewRequestDto;
+import com.sparta.blackwhitedeliverydriver.review.dto.ReviewResponseDto;
+import com.sparta.blackwhitedeliverydriver.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.review.entity.Review;
+import com.sparta.blackwhitedeliverydriver.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.review.exception.ReviewExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.store.exception.StoreExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.review.repository.ReviewRepository;
+import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
 import jakarta.validation.Valid;
 import java.time.LocalDateTime;
 import java.util.UUID;
